# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vorea Studio API",
    "version": "2026-03-19",
    "summary": "API backend completa de Vorea Studio (inventario generado automáticamente).",
    "description": "Especificación OpenAPI generada a partir de rutas reales del backend. PayPal es la pasarela operativa actual."
  },
  "servers": [
    {
      "url": "http://localhost:5173",
      "description": "Frontend local"
    },
    {
      "url": "http://localhost:3001",
      "description": "API server local"
    }
  ],
...
  "paths": {
...
    "/api/internal/news/cleanup": {
      "post": {
        "operationId": "post_api_internal_news_cleanup",
        "tags": [
          "Internal News"
        ],
        "summary": "Crear/ejecutar /api/internal/news/cleanup",
        "description": "Endpoint real detectado en `server/news-routes.ts:70`.\n\nPolítica de acceso: **public**.\n\nDependencias primarias: ninguna explícita.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Respuesta exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccess"
                }
              }
            }
          },
          "401": {
            "description": "Autenticación requerida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericError"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericError"
                }
              }
            }
          }
        },
        "x-vorea-meta": {
          "authPolicy": "public",
          "role": "public",
          "status": "active",
          "dependencies": [],
          "source": "server/news-routes.ts:70"
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              },
              "example": {
                "note": "Consultar contrato detallado en matriz operativa."
              }
            }
          }
        }
      }
    },
    "/api/internal/news/ingest": {
      "post": {
        "operationId": "post_api_internal_news_ingest",
        "tags": [
          "Internal News"
        ],
        "summary": "Crear/ejecutar /api/internal/news/ingest",
        "description": "Endpoint real detectado en `server/news-routes.ts:51`.\n\nPolítica de acceso: **public**.\n\nDependencias primarias: ninguna explícita.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Respuesta exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccess"
                }
              }
            }
          },
          "401": {
            "description": "Autenticación requerida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericError"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericError"
                }
              }
            }
          }
        },
        "x-vorea-meta": {
          "authPolicy": "public",
          "role": "public",
          "status": "active",
          "dependencies": [],
          "source": "server/news-routes.ts:51"
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              },
              "example": {
                "note": "Consultar contrato detallado en matriz operativa."
              }
            }
          }
        }
      }
    },
    "/api/news": {
      "get": {
        "operationId": "get_api_news",
        "tags": [
          "News"
        ],
        "summary": "Listar /api/news",
        "description": "Endpoint real detectado en `server/news-routes.ts:22`.\n\nPolítica de acceso: **public**.\n\nDependencias primarias: ninguna explícita.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Respuesta exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccess"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericError"
                }
              }
            }
          }
        },
        "x-vorea-meta": {
          "authPolicy": "public",
          "role": "public",
          "status": "active",
          "dependencies": [],
          "source": "server/news-routes.ts:22"
        }
      }
    },
    "/api/news/{slug}": {
      "get": {
        "operationId": "get_api_news_slug",
        "tags": [
          "News"
        ],
        "summary": "Obtener /api/news/:slug",
        "description": "Endpoint real detectado en `server/news-routes.ts:38`.\n\nPolítica de acceso: **public**.\n\nDependencias primarias: ninguna explícita.",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Parámetro de ruta: slug"
          }
        ],
        "responses": {
          "200": {
            "description": "Respuesta exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccess"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericError"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericError"
                }
              }
            }
          }
        },
        "x-vorea-meta": {
          "authPolicy": "public",
          "role": "public",
          "status": "active",
          "dependencies": [],
          "source": "server/news-routes.ts:38"
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "GenericSuccess": {
        "type": "object",
        "additionalProperties": true,
        "description": "Respuesta JSON exitosa (forma varía por endpoint)."
      },
      "GenericError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": true,
        "description": "Respuesta JSON de error."
      }
    }
  },
...
}